openclaw-man-server/pyproject.toml
[project]
name = "openclaw-man-server"
version = "0.1.0"
description = "Custom OpenClaw channel integration for ManServer"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "websockets>=12.0",
    "pyyaml>=6.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.0",
    "pymysql>=1.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.27.0",
    "python-dotenv>=1.2.1",
]

[project.scripts]
man-server = "openclaw_man_server.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
